Some quick polishing before Midterm Demo
Dog rocks are now dog houses, The main menu now has a background. the
default sprite for the monster is now the side view. Tall dog is now a
Mastiff for not having a trees sake. Still need to do demo level, place
colliders on things and fine tune them to work. Know how  just need to
do it now